# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateAnalysis": {
        "properties": {
          "breakdown": {
            "items": {
              "$ref": "#/components/schemas/ScoreLine"
            },
            "title": "Breakdown",
            "type": "array"
          },
          "codec": {
            "title": "Codec",
            "type": "string"
          },
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "exclusion_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclusion Reason"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "manual": {
            "title": "Manual",
            "type": "boolean"
          },
          "manual_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Manual Reasons",
            "type": "array"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "profile_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Label"
          },
          "ranking": {
            "title": "Ranking",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "recommendation": {
            "title": "Recommendation",
            "type": "string"
          },
          "resolution": {
            "title": "Resolution",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "integer"
          }
        },
        "required": [
          "ranking",
          "score",
          "breakdown",
          "tier",
          "group",
          "resolution",
          "codec",
          "source",
          "recommendation",
          "reason",
          "manual",
          "manual_reasons",
          "eligible",
          "exclusion_reason",
          "profile_id",
          "profile_label"
        ],
        "title": "CandidateAnalysis",
        "type": "object"
      },
      "ConfigPayload": {
        "properties": {
          "candidate_limit": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Limit"
          },
          "candidate_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Policy"
          }
        },
        "title": "ConfigPayload",
        "type": "object"
      },
      "ConnectionStatus": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "providers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ProviderCheck"
            },
            "title": "Providers",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "message",
          "version",
          "providers"
        ],
        "title": "ConnectionStatus",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProviderCheck": {
        "additionalProperties": true,
        "properties": {
          "configured": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configured"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ok"
          }
        },
        "title": "ProviderCheck",
        "type": "object"
      },
      "PublicSettings": {
        "description": "运行设置：密钥只返回是否已配置（字段本身为空或掩码）。",
        "properties": {
          "access_token_required": {
            "title": "Access Token Required",
            "type": "boolean"
          },
          "access_token_strength": {
            "enum": [
              "missing",
              "weak",
              "strong"
            ],
            "title": "Access Token Strength",
            "type": "string"
          },
          "access_token_strength_enforced": {
            "title": "Access Token Strength Enforced",
            "type": "boolean"
          },
          "ai_api_key": {
            "title": "Ai Api Key",
            "type": "string"
          },
          "ai_api_key_configured": {
            "title": "Ai Api Key Configured",
            "type": "boolean"
          },
          "ai_base_url": {
            "title": "Ai Base Url",
            "type": "string"
          },
          "ai_model": {
            "title": "Ai Model",
            "type": "string"
          },
          "cookiecloud_key": {
            "title": "Cookiecloud Key",
            "type": "string"
          },
          "cookiecloud_key_configured": {
            "title": "Cookiecloud Key Configured",
            "type": "boolean"
          },
          "cookiecloud_password": {
            "title": "Cookiecloud Password",
            "type": "string"
          },
          "cookiecloud_password_configured": {
            "title": "Cookiecloud Password Configured",
            "type": "boolean"
          },
          "cookiecloud_url": {
            "title": "Cookiecloud Url",
            "type": "string"
          },
          "cookiecloud_url_configured": {
            "title": "Cookiecloud Url Configured",
            "type": "boolean"
          },
          "dashboard_random_posters": {
            "title": "Dashboard Random Posters",
            "type": "boolean"
          },
          "emby_api_key": {
            "title": "Emby Api Key",
            "type": "string"
          },
          "emby_api_key_configured": {
            "title": "Emby Api Key Configured",
            "type": "boolean"
          },
          "emby_base_url": {
            "title": "Emby Base Url",
            "type": "string"
          },
          "fanart_api_key": {
            "title": "Fanart Api Key",
            "type": "string"
          },
          "fanart_api_key_configured": {
            "title": "Fanart Api Key Configured",
            "type": "boolean"
          },
          "mdblist_api_key": {
            "title": "Mdblist Api Key",
            "type": "string"
          },
          "mdblist_api_key_configured": {
            "title": "Mdblist Api Key Configured",
            "type": "boolean"
          },
          "mp_api_key": {
            "title": "Mp Api Key",
            "type": "string"
          },
          "mp_api_key_configured": {
            "title": "Mp Api Key Configured",
            "type": "boolean"
          },
          "mp_base_url": {
            "title": "Mp Base Url",
            "type": "string"
          },
          "mp_timeout_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "title": "Mp Timeout Seconds"
          },
          "outbound_proxy_configured": {
            "title": "Outbound Proxy Configured",
            "type": "boolean"
          },
          "outbound_proxy_url": {
            "title": "Outbound Proxy Url",
            "type": "string"
          },
          "outbound_proxy_url_configured": {
            "title": "Outbound Proxy Url Configured",
            "type": "boolean"
          },
          "pt_proxy_enabled": {
            "title": "Pt Proxy Enabled",
            "type": "boolean"
          },
          "tmdb_api_key": {
            "title": "Tmdb Api Key",
            "type": "string"
          },
          "tmdb_api_key_configured": {
            "title": "Tmdb Api Key Configured",
            "type": "boolean"
          },
          "tmdb_language": {
            "title": "Tmdb Language",
            "type": "string"
          },
          "tmdb_proxy_enabled": {
            "title": "Tmdb Proxy Enabled",
            "type": "boolean"
          },
          "tr_base_url": {
            "title": "Tr Base Url",
            "type": "string"
          },
          "tr_password": {
            "title": "Tr Password",
            "type": "string"
          },
          "tr_password_configured": {
            "title": "Tr Password Configured",
            "type": "boolean"
          },
          "tr_username": {
            "title": "Tr Username",
            "type": "string"
          },
          "tr_username_configured": {
            "title": "Tr Username Configured",
            "type": "boolean"
          }
        },
        "required": [
          "mp_base_url",
          "mp_api_key",
          "mp_api_key_configured",
          "mp_timeout_seconds",
          "emby_base_url",
          "emby_api_key",
          "emby_api_key_configured",
          "tmdb_api_key",
          "tmdb_api_key_configured",
          "tmdb_language",
          "fanart_api_key",
          "fanart_api_key_configured",
          "mdblist_api_key",
          "mdblist_api_key_configured",
          "cookiecloud_url",
          "cookiecloud_url_configured",
          "cookiecloud_key",
          "cookiecloud_key_configured",
          "cookiecloud_password",
          "cookiecloud_password_configured",
          "outbound_proxy_configured",
          "outbound_proxy_url_configured",
          "outbound_proxy_url",
          "pt_proxy_enabled",
          "tmdb_proxy_enabled",
          "ai_base_url",
          "ai_api_key",
          "ai_api_key_configured",
          "ai_model",
          "tr_base_url",
          "tr_username",
          "tr_username_configured",
          "tr_password",
          "tr_password_configured",
          "dashboard_random_posters",
          "access_token_required",
          "access_token_strength",
          "access_token_strength_enforced"
        ],
        "title": "PublicSettings",
        "type": "object"
      },
...
      "ScoreLine": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "label",
          "score"
        ],
        "title": "ScoreLine",
        "type": "object"
      },
      "ScorePreviewPayload": {
        "properties": {
          "candidate_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Policy"
          },
          "seeders": {
            "default": 0,
            "minimum": 0.0,
            "title": "Seeders",
            "type": "integer"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "volume_factor": {
            "default": 1,
            "minimum": 0.0,
            "title": "Volume Factor",
            "type": "number"
          }
        },
        "required": [
          "title"
        ],
        "title": "ScorePreviewPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AutoList"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/config": {
      "get": {
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Get Config Api Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config"
      },
      "put": {
        "operationId": "put_config_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Put Config Api Config Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Config"
      }
    },
...
    "/api/config/score-preview": {
      "post": {
        "operationId": "score_preview_api_config_score_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScorePreviewPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateAnalysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score Preview"
      }
    },
    "/api/connection": {
      "get": {
        "operationId": "connection_api_connection_get",
        "parameters": [
          {
            "in": "query",
            "name": "force_refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connection"
      }
    },
...
    "/api/settings": {
      "get": {
        "operationId": "get_runtime_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Runtime Settings"
      },
      "put": {
        "operationId": "put_runtime_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeSettingsPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Runtime Settings"
      }
    },
    "/api/settings/test": {
      "post": {
        "operationId": "test_runtime_settings_api_settings_test_post",
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "tmdb",
                    "fanart",
                    "emby",
                    "transmission",
                    "moviepilot",
                    "cookiecloud"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/ProviderCheck"
                  },
                  "title": "Response Test Runtime Settings Api Settings Test Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Runtime Settings"
      }
    },
...
  }
}